requirements.txt
flet>=0.2.4
google-cloud-aiplatform>=1.25
shapely<2.0.0
